First actual submission. Let's go!
If you are someone who finds this on accident, I would just like to preface that this is a slightly over-the-top version of what should be a paper requirement in our school. Our teacher asking for a stamp card where we can record our recitation points. The critical part is that it could be from paper OR digital. I opted for the one that only needed a screen and a keyboard.